requirements.txt
astroid==1.6.3
backcall==0.1.0
click==6.7
decorator==4.3.0
Django==2.0.5
django-environ==0.4.5
django-extensions==2.0.7
django-money==0.14.3
django-salmonella==1.2
django-taggit==0.22.2
django-widget-tweaks==1.4.2
envs==1.2.6
gunicorn==19.9.0
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py-moneyed==0.7.0
Pygments==2.2.0
pylint==1.8.4
pytz==2018.4
simplegeneric==0.8.1
six==1.11.0
terminaltables==3.1.0
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==4.1
wrapt==1.10.11
